pyproject.toml
[project]
name = "frel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "joblib>=1.5.2",
    "lightgbm>=4.6.0",
    "mapie>=0.9.0",
    "matplotlib>=3.10.6",
    "pandas>=2.3.3",
    "pathlib>=1.0.1",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.3.0",
    # NOTE: tensorflow NOT managed by UV due to Windows CUDA package incompatibilities
    # Install separately with: uv run pip install tensorflow==2.16.2
    # See UV_VS_PIP_FOR_TENSORFLOW.md for details
    "tqdm>=4.67.1",
    "yfinance>=0.2.50",
]
